+++ packages/sveltefirets/src/lib/interfaces/index.ts
@@ -1,8 +1,11 @@
-import type { IBaseUser } from './user.interface';
+import type { IBaseUser, SignInMethods } from './user.interface';
 import type { IFirestoreMetaData, IFirestoreMetaDataAbbreviated } from './firestore-metadata.interface';
+import type { AuthResult } from './auth-result.interface';
 
 export type {
   IBaseUser,
   IFirestoreMetaData,
   IFirestoreMetaDataAbbreviated,
+  AuthResult,
+  SignInMethods
 };